deploy setting

+++ app/controllers/admin/dusts_controller.rb
@@ -1,4 +1,5 @@
 class Admin::DustsController < ApplicationController
+  before_action :authenticate_admin!
   def new
     @dust = Dust.new
   end
@@ -7,10 +8,10 @@ class Admin::DustsController < ApplicationController
     @dust = Dust.new(dust_params)
     @dust.admin_id = current_admin.id
     if @dust.save
-      redirect_to dusts_path(@dust.id)
+      redirect_to admin_dusts_path(@dust.id)
     else
-      @dusts = Dust.all
-      render :index
+      @dusts = Dust.new
+      render :new
     end
   end
 
@@ -21,7 +22,6 @@ class Admin::DustsController < ApplicationController
 
   def show
     @dust = Dust.find(params[:id])
-    @dust_comment = DustComment.new
   end
 
   def edit
@@ -29,7 +29,7 @@ class Admin::DustsController < ApplicationController
     if @dust.admin == current_admin
       render :edit
     else
-      redirect_to dusts_path
+      redirect_to admin_dusts_path
     end
   end
 
@@ -37,7 +37,7 @@ class Admin::DustsController < ApplicationController
     @dust = Dust.find(params[:id])
     if @dust.update(dust_params)
       flash[:notice] = "You have updated user successfully."
-      redirect_to dust_path(@dust.id)
+      redirect_to admin_dust_path(@dust.id)
     else
       render :edit
     end
@@ -45,31 +45,29 @@ class Admin::DustsController < ApplicationController
 
   def destroy
     dust = Dust.find(params[:id])
-    if dust.admin
-      dust.destroy
-    end
-    redirect_to '/dusts'
+    dust.destroy if Dust
+    redirect_to '/admin/dusts'
   end
 
   def favorite_create
-    dust = Dust.find(params[:dust_id])
-    if dust.admin.== current_admin
+    dust = Dust.find(params[:admin_dust_id])
+    if dust.admin.!= current_admin
       dust.favorites.create(admin_id: current_admin.id)
     end
     redirect_back fallback_location: root_path
   end
 
   def favorite_destroy
-    dust = Dust.find(params[:dust_id])
-    if dust.admin.== current_admin
-      dust.favorites.find_by(admin_id: current_admin.id).destroy
+    @dust = Dust.find(params[:admin_dust_id])
+    if @dust.admin.!= current_admin
+      @dust.favorites.find_by(admin_id: current_admin.id).destroy
     end
-    redirect_back fallback_location: root_path
   end
 
   private
 
-  def post_params
-    params.require(:dust).permit(:title, :body)
+  def dust_params
+    params.require(:dust).permit(:title, :body, :is_published_flag)
   end
+
 end
