+++ app/controllers/admin/posts_controller.rb
@@ -1,4 +1,5 @@
 class Admin::PostsController < ApplicationController
+  before_action :authenticate_admin!
   def new
     @post = Post.new
   end
@@ -7,15 +8,15 @@ class Admin::PostsController < ApplicationController
     @post = Post.new(post_params)
     @post.admin_id = current_admin.id
     if @post.save
-      redirect_to posts_path(@post.id)
+      redirect_to admin_post_path(@post.id)
     else
-      @posts = Post.all
+      @posts = Post.published
       render :index
     end
   end
 
   def index
-    @posts = Post.all
+    @posts = Post.published
     @post = Post.new
   end
 
@@ -29,7 +30,7 @@ class Admin::PostsController < ApplicationController
     if @post.admin == current_admin
       render :edit
     else
-      redirect_to posts_path
+      redirect_to admin_posts_path
     end
   end
 
@@ -37,7 +38,7 @@ class Admin::PostsController < ApplicationController
     @post = Post.find(params[:id])
     if @post.update(post_params)
       flash[:notice] = "You have updated user successfully."
-      redirect_to post_path(@post.id)
+      redirect_to admin_post_path(@post.id)
     else
       render :edit
     end
@@ -45,31 +46,29 @@ class Admin::PostsController < ApplicationController
 
   def destroy
     post = Post.find(params[:id])
-    if post.admin == current_admin
-      post.destroy
-    end
-    redirect_to '/posts'
+    post.destroy if post
+    redirect_to '/admin/posts'
   end
 
   def favorite_create
-    post = Post.find(params[:post_id])
-    if post.admin.== current_admin
+    post = Post.find(params[:admin_post_id])
+    if post.admin.!= current_admin
       post.favorites.create(admin_id: current_admin.id)
     end
     redirect_back fallback_location: root_path
   end
 
   def favorite_destroy
-    post = Post.find(params[:post_id])
-    if post.admin.== current_admin
-      post.favorites.find_by(admin_id: current_admin.id).destroy
+    @post = Post.find(params[:admin_post_id])
+    if @post.admin.!= current_admin
+      @post.favorites.find_by(admin_id: current_admin.id).destroy
     end
-    redirect_back fallback_location: root_path
   end
 
   private
 
   def post_params
-    params.require(:post).permit(:title, :body)
+    params.require(:post).permit(:title, :body, :is_published_flag)
   end
+  
 end
